Add comment to explain small differences in results from durations parsed with power_and_time_per_pulse and  rise_and_fall_times.
power_and_time_per_pulse returns time durations in the sampling frame of the power timestamps. Hence the list of times this function returns will not exactly match with the difference between rise timestamp and fall timestamp of the of the GPIO pin.

rise_and_fall_times returns the actual rise and fall timestamps of the pins.

The differences are smaller than 1e-4 per timestamp, and may depend on the accuracy of synchronization between power and timestamps